Add hardware_data_flow_diagram.png
I had already added this image to the repo readme via drag-and-drop, as a stackoverflow page had said it would work.
And it did for a couple of days until GitHub presumably deleted the temporary file!
Just use a image file in the repo instead.

Also added a .gitignore so I can keep the word document version of the report in the same place on my local machine, just for convenience if I need to pull any other images out of it. It's easier to extract them from the .docx than the .pdf.